Added a page display to the View
Page display was added to go along with the feed display. The feed display is available to the public (so Zapier can communicate with it). The page display is only accessible by a user with the administrator role. The page contains an HTML list of content with the Needs Review status.